+++ src/pages/Community.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import { RxBookmarkFilled, RxBookmark } from "react-icons/rx";
 import { Link, useNavigate } from "react-router-dom";
 
 import { Arrow } from "@egjs/flicking-plugins";
@@ -7,8 +6,7 @@ import Flicking, { ViewportSlot } from "@egjs/react-flicking";
 import { supabase } from "api/supabase";
 import { DateConvertor } from "components/date";
 import { PostPagination } from "components/pagination";
-import { usePostsBookmark } from "hooks";
-import { useAuthStore } from "store";
+import { PostBookmark } from "components/postBookmark";
 import { type Tables } from "types/supabase";
 import "@egjs/flicking-plugins/dist/arrow.css";
 import "@egjs/react-flicking/dist/flicking.css";
@@ -17,17 +15,11 @@ export const POSTS_PER_PAGE = 4;
 export const storageUrl = process.env.REACT_APP_SUPABASE_STORAGE_URL as string;
 const plugins = [new Arrow()];
 
-interface FetchPostBookmark {
-  postId: string;
-  userId: string;
-}
-
 export const Community = () => {
   const navigate = useNavigate();
   const [postList, setPostList] = useState<Array<Tables<"POSTS", "Row">>>([]);
   const [currentPage, setCurrentPage] = useState<number>(1);
   const [selectedOption, setSelectedOption] = useState<string>("");
-  const [isPostBookmarkedData, setIsPostBookmarkedData] = useState<FetchPostBookmark[]>();
 
   const handleOptionChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
     setSelectedOption(event.target.value);
@@ -71,16 +63,6 @@ export const Community = () => {
   const indexOfLastPost = currentPage * POSTS_PER_PAGE;
   const indexOfFirstPost = indexOfLastPost - POSTS_PER_PAGE;
   const currentFilteredPosts = filteredPosts.slice(indexOfFirstPost, indexOfLastPost);
-
-  const { currentSession } = useAuthStore();
-
-  const { postBookmarkResponse, addBookmarkMutation, deleteBookmarkMutation } = usePostsBookmark();
-  const { data: currentBookmarkData } = postBookmarkResponse;
-
-  useEffect(() => {
-    if (currentBookmarkData == null) return;
-    setIsPostBookmarkedData(currentBookmarkData);
-  }, [currentBookmarkData, currentSession]);
 
   return (
     <div className="w-[1200px] mx-auto">
@@ -140,10 +122,6 @@ export const Community = () => {
             게시물 작성
           </Link>
           {currentFilteredPosts.map((post) => {
-            let isPostBookmark: FetchPostBookmark | null | undefined = null;
-            if (isPostBookmarkedData !== undefined) {
-              isPostBookmark = isPostBookmarkedData.find((bookmarkItem) => bookmarkItem.postId === post.id);
-            }
             return (
               <div key={post.id} className="border-b border-[#dddddd] py-5 my-5">
                 <div onClick={goDetailPage} className="flex justify-between gap-5 cursor-pointer">
@@ -169,33 +147,7 @@ export const Community = () => {
                   <p>
                     <DateConvertor datetime={post.created_at} type="dotDate" />
                   </p>
-                  {isPostBookmark !== undefined ? (
-                    <RxBookmark
-                      className="text-[25px] cursor-pointer"
-                      onClick={async () => {
-                        if (currentSession === null) {
-                          alert("북마크 기능은 로그인 후 이용가능합니다.");
-                          return;
-                        }
-                        if (isPostBookmark == null) return;
-                        deleteBookmarkMutation.mutate({
-                          postId: isPostBookmark.postId,
-                          userId: currentSession.user.id,
-                        });
-                      }}
-                    />
-                  ) : (
-                    <RxBookmarkFilled
-                      className="text-[25px] cursor-pointer"
-                      onClick={async () => {
-                        if (currentSession === null) {
-                          alert("북마크 기능은 로그인 후 이용가능합니다.");
-                          return;
-                        }
-                        addBookmarkMutation.mutate({ postId: post.id, userId: currentSession.user.id });
-                      }}
-                    />
-                  )}
+                  <PostBookmark postId={post.id} />
                 </div>
               </div>
             );
